backend/fenix_evolucao/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
cryptography>=44.0
httpx>=0.28
sentence-transformers>=3.3.0
langchain>=0.3.0
langchain-community>=0.3.0
chromadb>=0.5.0
